refactor(matrices): add MutableRepMatrix
